+++ cmd/aitaskbuilder/batches.go
@@ -18,6 +18,7 @@ func NewBatchesCommand(client client.API, w io.Writer) *cobra.Command {
 	cmd.AddCommand(
 		NewBatchCreateCommand(client, w),
 		NewBatchInstructionsCommand(client, w),
+		NewBatchSetupCommand(client, w),
 		NewGetBatchCommand(client, w),
 		NewGetBatchStatusCommand(client, w),
 		NewGetBatchesListCommand(client, w),